Refactor ApplicationBaseComponent and nest tabs in ApplicationTabsComponent
