--- src/app/features/applications/application/application-base.component.scss
@@ -1,28 +1,0 @@
-.summary {
-  display: flex;
-  &__nav {
-    margin-top: 20px;
-    margin-bottom: 20px;
-  }
-  &__form {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    flex-wrap: wrap;
-    div {
-      flex: 1;
-      display: flex;
-      flex-direction: row;
-    }
-  }
-  &__card {
-    flex: 1;
-  }
-  mat-chip-list {
-    mat-chip {
-      margin-right: 7px;
-      margin-bottom: 5px;
-      margin-left: 0;
-    }
-  }
-}
