+++ src/app/features/applications/application/application-tabs-base/application-tabs-base.component.html
@@ -1,0 +1,136 @@
+<app-page-header>
+  <h1>{{ (applicationService.application$ | async)?.app.entity.name }} </h1>
+  <div class="page-header-right">
+    <span *ngIf="isEditSummary">
+      <button mat-icon-button mat-button (click)="saveEdits()" [disabled]="!summaryForm.valid">
+        <mat-icon>done</mat-icon>
+      </button>
+      <button mat-icon-button name="cancel" mat-button (click)="endEdit()">
+        <mat-icon>clear</mat-icon>
+      </button>
+    </span>
+    <span *ngIf="!isEditSummary">
+      <button mat-icon-button name="delete" [disabled]="summaryDataChanging$ | async" (click)="deleteApplication()">
+        <mat-icon>delete</mat-icon>
+      </button>
+      <button mat-icon-button name="edit" *ngIf="(this.applicationService.applicationState$ | async)" [disabled]="summaryDataChanging$ | async"
+        (click)="startEdit()">
+        <mat-icon>edit</mat-icon>
+      </button>
+      <span>
+        <button mat-icon-button name="stop" *ngIf="(applicationService.application$ | async)?.app.entity.state === 'STARTED'" [disabled]="summaryDataChanging$ | async"
+          (click)="stopApplication()">
+          <mat-icon>stop</mat-icon>
+        </button>
+        <button mat-icon-button name="start" *ngIf="(applicationService.application$ | async)?.app.entity.state === 'STOPPED'" [disabled]="summaryDataChanging$ | async"
+          (click)="startApplication()">
+          <mat-icon>play_arrow</mat-icon>
+        </button>
+      </span>
+    </span>
+  </div>
+</app-page-header>
+
+<app-loading-page [isLoading]="isFetching$" [text]="'Retrieving your application'">
+  <div class="summary">
+    <app-details-card class="summary__form" title="Summary" [busy]="summaryDataChanging$ | async">
+      <form class="summary__form" #summaryForm="ngForm">
+        <div>
+          <mat-list role='list'>
+            <h3 mat-subheader>Application State</h3>
+            <mat-list-item role="listitem">
+              <mat-icon mat-list-icon></mat-icon>
+              <div mat-line>
+                <app-application-state [state]="applicationService.applicationState$ | async"></app-application-state>
+              </div>
+            </mat-list-item>
+            <mat-list-item role="listitem">
+              <mat-icon mat-list-icon></mat-icon>
+              <div mat-line>
+                <mat-chip-list aria-orientation="vertical">
+                  <mat-chip>App: {{(applicationService.application$ | async)?.app.entity.state}}</mat-chip>
+                  <mat-chip>Package: {{(applicationService.application$| async)?.app.entity.package_state}}</mat-chip>
+                </mat-chip-list>
+              </div>
+            </mat-list-item>
+          </mat-list>
+          <mat-list role='list'>
+            <h3 mat-subheader>Information</h3>
+            <mat-list-item role="listitem">
+              <mat-icon mat-list-icon>timeline</mat-icon>
+              <div mat-line>Running Instances</div>
+              <div mat-line>
+                <app-editable-display-value [edit]="isEditSummary">
+                  <div class="display-input__display">
+                    <app-display-value>
+                      {{(applicationService.appSummary$| async)?.metadata?.running_instances || 0}} / {{ (applicationService.application$ | async)?.app.entity?.instances }}
+                    </app-display-value>
+                  </div>
+                  <div class="display-input__input">
+                    <mat-form-field>
+                      <input matInput name="instances" type="number" placeholder="Required No. Instances" [(ngModel)]="appEdits.instances" required pattern="[0-9]+" />
+                    </mat-form-field>
+                  </div>
+                </app-editable-display-value>
+              </div>
+            </mat-list-item>
+            <mat-list-item role="listitem">
+              <mat-icon mat-list-icon>developer_board</mat-icon>
+              <div mat-line>Services</div>
+              <div mat-line>{{(applicationService.appSummary$| async)?.metadata?.services?.length || '-'}}</div>
+            </mat-list-item>
+          </mat-list>
+        </div>
+        <div>
+          <mat-list role='list'>
+            <h3 mat-subheader>Usage</h3>
+            <mat-list-item role="listitem">
+              <mat-icon mat-list-icon>memory</mat-icon>
+              <div mat-line>Memory</div>
+              <div mat-line>
+                <app-editable-display-value [edit]="isEditSummary" [value]="(applicationService.application$| async)?.app.entity.memory | mbToHumanSize">
+                  <div class="display-input__input">
+                    <mat-form-field>
+                      <input matInput name="appMemory" type="number" placeholder="Memory" [(ngModel)]="appEdits.memory" required pattern="[0-9]+" />
+                    </mat-form-field>
+                  </div>
+                </app-editable-display-value>
+              </div>
+            </mat-list-item>
+            <mat-list-item role="listitem">
+              <mat-icon mat-list-icon>storage</mat-icon>
+              <div mat-line>Disk quota</div>
+              <div mat-line>{{(applicationService.application$| async)?.app.entity.disk_quota | mbToHumanSize}}</div>
+            </mat-list-item>
+          </mat-list>
+          <mat-list role='list'>
+            <mat-list-item role="listitem">
+              <mat-icon mat-list-icon>developer_mode</mat-icon>
+              <div mat-line>SSH Enabled</div>
+              <div mat-line>
+                <app-editable-display-value [edit]="isEditSummary">
+                  <div class="display-input__display">
+                    <app-display-value>
+                      <mat-slide-toggle name="appSsh" [ngModel]="(applicationService.application$| async)?.app.entity.enable_ssh" [disabled]="true"></mat-slide-toggle>
+                    </app-display-value>
+                  </div>
+                  <div class="display-input__input">
+                    <app-display-value>
+                      <mat-slide-toggle name="appSsh" [ngModel]="appEdits.enable_ssh"></mat-slide-toggle>
+                    </app-display-value>
+                  </div>
+                </app-editable-display-value>
+              </div>
+            </mat-list-item>
+          </mat-list>
+        </div>
+      </form>
+    </app-details-card>
+  </div>
+  <nav mat-tab-nav-bar class="summary__nav">
+    <a mat-tab-link *ngFor="let tabLink of tabLinks" [routerLink]="[tabLink.link]" routerLinkActive="active-link" #rla="routerLinkActive" [active]="rla.isActive">
+      {{tabLink.label}}
+    </a>
+  </nav>
+  <router-outlet></router-outlet>
+</app-loading-page>
