+++ src/app/features/applications/application/application-tabs-base/tabs/build-tab/application-env-vars.service.spec.ts
