+++ src/app/features/applications/application/application-tabs-base/tabs/build-tab/build-tab.component.scss
