+++ src/app/features/applications/application/application-tabs-base/tabs/build-tab/build-tab.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { Observable } from 'rxjs/Rx';
 
-import { ApplicationData, ApplicationService } from '../../application.service';
-import { AppMetadataInfo } from '../../../../store/types/app-metadata.types';
+import { ApplicationData, ApplicationService } from '../../../../application.service';
+import { AppMetadataInfo } from '../../../../../../store/types/app-metadata.types';
 
 @Component({
   selector: 'app-build-tab',
