+++ src/app/features/applications/application/application-tabs-base/tabs/build-tab/view-buildpack/view-buildpack.component.ts
