+++ src/app/features/applications/application/application-tabs-base/tabs/events-tab/events-tab.component.scss
