+++ src/app/features/applications/application/application-tabs-base/tabs/log-stream-tab/log-stream-tab.component.html
