+++ src/app/features/applications/application/application-tabs-base/tabs/ssh-tab/ssh-tab.component.ts
