+++ src/app/features/applications/application/application-tabs-base/tabs/variables-tab/variables-tab.component.spec.ts
