+++ src/app/shared/components/table/custom-cells/table-cell-app-status/table-cell-app-status.component.ts
@@ -4,9 +4,6 @@ import { Store } from '@ngrx/store';
 import { AppState } from '../../../../../store/app-state';
 import { ActivatedRoute } from '@angular/router';
 import { EntityService } from '../../../../../core/entity-service';
-import {
-  ApplicationEnvVarsService,
-} from '../../../../../features/applications/application/build-tab/application-env-vars.service';
 import { ApplicationService } from '../../../../../features/applications/application.service';
 import { ApplicationSchema, GetApplication } from '../../../../../store/actions/application.actions';
 import { Component, Input, OnDestroy, OnInit } from '@angular/core';
